SignIn.jsをindex.jsに移植
