+++ frontend/src/pages/index.js
@@ -1,40 +1,90 @@
-import Head from "next/head";
-import styles from "@/styles/Home.module.css";
-import AccountButton from "../../Login/account-button";
-import LoginButton from "../../Login/login-button";
-import LoginForm from "../../Login/loginform";
+"use client";
+
+import React, { useState } from 'react';
+import { auth } from '../firebase'; // パスを適切に調整
+import { signInWithEmailAndPassword } from 'firebase/auth';
+import { useRouter } from 'next/router';
+import s from '../styles/SignIn.module.css';
+import {Inter} from "next/font/google";
+import styles from "@/styles/loginform.module.css";
 import Comment from "../../Login/comment";
+import AccountButton from "../../Login/account-button";
+
+const inter = Inter({ subsets: ["latin"] });
 
 
+const SignIn = () => {
+    const [email, setEmail] = useState('');
+    const [password, setPassword] = useState('');
+    const [error, setError] = useState(null);
+    const router = useRouter(); // useRouter フックを使用
+    const [isRevealPassword, setIsRevealPassword] = useState(false);
 
-export default function Home() {
+    const handleSignIn = async (event) => {
+        event.preventDefault();
+        try {
+            await signInWithEmailAndPassword(auth, email, password);
+            alert('ログイン成功');
+            router.push('/home'); // ログイン成功後のリダイレクト
+        } catch (error) {
+            setError(error.message);
+        }
+    };
+
+    const navigateToSignUp = () => {
+        router.push('/SignUp'); // SignUp.js への遷移
+    };
+
+    const togglePassword = () => {
+        setIsRevealPassword((prevState) => !prevState);
+    };
+
     return (
-        <>
-            <Head>
-                <title>Create Next App</title>
-                <meta name="description" content="Generated by create next app" />
-                <meta name="viewport" content="width=device-width, initial-scale=1" />
-                <link rel="icon" href="/favicon.ico" />
-            </Head>
-            <main className={`${styles.main}`}>
-                <div className={styles.divdiv}>
-                    {/*--ログインフォーム--*/}
-                    <LoginForm />
+        <main className={`${s.main}`}>
+            <h1 className={s.heading}>Sign In</h1>
+            <div className={s.divdiv}>
+                <form onSubmit={handleSignIn} className={s.form}>
 
-                    {/*--新規登録、ログインボタン--*/}
-                    <LoginButton />
+                    <p className={s.email_comment}>Email</p>
+                    <input
+                        type="email"
+                        value={email}
+                        onChange={(e) => setEmail(e.target.value)}
+                        placeholder="email@icloud.com"
+                        className={s.mail_address}
+                    />
 
-                    {/*--文章--*/}
-                    <Comment />
+                    <p className={s.email_comment2}>Password</p>
+                    <input
+                        type={isRevealPassword ? "text" : "password"}
+                        value={password}
+                        onChange={(e) => setPassword(e.target.value)}
+                        placeholder="123456"
+                        className={s.password_text}/>
+                    <span
+                        onClick={togglePassword}
+                        role="presentation"
+                        className={styles.password_parts}>
+            {isRevealPassword ? (
+                <i className="fas fa-eye"/>
+            ) : (
+                <i className="fas fa-eye-slash"/>
+            )}
+        </span>
+                    <div className={s.buttonGroup}>
+                        <button type="button" className={s.touroku_button} onClick={navigateToSignUp}>新規登録</button>
+                        <button type="submit" className={s.login_button}>Sign In</button>
+                    </div>
+                </form>
+                {error && <p className={s.error}>{error}</p>}
 
-                    {/*--google--*/}
-                    <AccountButton />
+                <Comment/>
 
-                </div>
-            </main>
-        </>
+                <AccountButton/>
+
+            </div>
+        </main>
     );
 };
 
-
-
+export default SignIn;
