+++ frontend/src/styles/SignIn.module.css
@@ -49,4 +49,125 @@
 
 .error {
     color: red;
+    position: relative;
+    left: 90px;
+    top: 270px;
+}
+
+
+.password_text {
+    width: 350px;
+    height: 49px;
+    left: 60px;
+    top: 140px;
+    padding: 25px;
+    position: absolute;
+    border-radius: 54px;
+    color: #8E8585;
+    font-size: 18px;
+    font-family: Inter;
+    font-weight: 400;
+    word-wrap: break-word;
+    letter-spacing: 1px;
+    border-color: white;
+    background-color: whitesmoke;
+    outline: none;
+}
+.password_text:focus {
+    border-color: cornflowerblue;
+}
+
+.password_parts {
+    position: absolute;
+    left: 360px;
+    top: 154px;
+    color: #747474;
+}
+
+.mail_address {
+    width: 350px;
+    height: 49px;
+    left: 60px;
+    top: 50px;
+    padding: 25px;
+    position: absolute;
+    color: #8E8585;
+    border-radius: 54px;
+    font-size: 18px;
+    font-family: Inter;
+    font-weight: 400;
+    word-wrap: break-word;
+    border-color: white;
+    background-color: whitesmoke;
+    outline: none;
+}
+
+.mail_address:focus {
+    border-color: cornflowerblue;
+}
+
+.email_comment {
+    transform: translateY(15px) translateX(-130px);
+    color: grey;
+}
+
+.email_comment2 {
+    transform: translateY(80px) translateX(-115px);
+    color: grey;
+}
+
+.login_button {
+    width: 350px;
+    height: 40px;
+    left: 60px;
+    top: 220px;
+    background-color: deepskyblue;
+    border-radius: 54px;
+    border-color: white;
+    position: absolute;
+    color: white;
+    font-size: 20px;
+    font-family: Inter;
+    font-weight: 400;
+    word-wrap: break-word;
+
+}
+
+.touroku_button {
+    width: 350px;
+    height: 40px;
+    left: 60px;
+    top: 270px;
+    position: absolute;
+    background-color: black;
+    color: white;
+    border-radius: 55px;
+    border-color: white;
+    font-size: 20px;
+    font-family: Inter;
+    font-weight: 400;
+    word-wrap: break-word;
+}
+
+.divdiv {
+    width: 460px;
+    height: 492px;
+    position: relative;
+    border: 1px solid black;
+    top: -180px;
+    padding: 10px;
+    margin: 10px;
+    background-color: white;
+    border-radius: 24px;
+    border-color: white;
+    box-shadow: 0px 187px 75px rgba(0, 0, 0, 0.01), 0px 105px 63px rgba(0, 0, 0, 0.05), 0px 47px 47px rgba(0, 0, 0, 0.09), 0px 12px 26px rgba(0, 0, 0, 0.1), 0px 0px 0px rgba(0, 0, 0, 0.1);
+}
+
+.main {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    align-items: center;
+    padding: 6rem;
+    min-height: 100vh;
 }
