+++ frontend/src/styles/comment.module.css
@@ -1,8 +1,7 @@
 .comment {
     width: 374px;
     font-size: 14px;
-    left: 20px;
     margin: 0 auto;
-    transform: translateY(280px);
+    transform: translateY(290px) translateX(15px);
 
 }
