+++ frontend/src/styles/loginform.module.css
@@ -58,3 +58,36 @@
     transform: translateY(80px) translateX(55px);
     color: grey;
 }
+
+.login_button {
+    width: 350px;
+    height: 40px;
+    left: 60px;
+    top: 220px;
+    background-color: deepskyblue;
+    border-radius: 54px;
+    border-color: white;
+    position: absolute;
+    color: white;
+    font-size: 20px;
+    font-family: Inter;
+    font-weight: 400;
+    word-wrap: break-word;
+
+}
+
+.touroku_button {
+    width: 350px;
+    height: 40px;
+    left: 60px;
+    top: 270px;
+    position: absolute;
+    background-color: black;
+    color: white;
+    border-radius: 55px;
+    border-color: white;
+    font-size: 20px;
+    font-family: Inter;
+    font-weight: 400;
+    word-wrap: break-word;
+}
